flask-api/requirements.txt
Flask>=2.0
python-jose>=3.3.0
requests>=2.25.0